Add context to conversations
