+++ galacticview_bot/response_structure.py
@@ -19,6 +19,6 @@ class ImageResponseStructure(BaseModel):
     )
 
 class SpaceResponseStructure(BaseModel):
-    title: str = Field(description="Title of the subject")
-    content: str = Field(description="Comprehensive summary of the subject matter, 100-150 words")
-    key_metrics: list[str] = Field(description="Bullet points of specific numbers/data found")
+    title: str = Field(description="A catchy title for the topic")
+    content: str = Field(description="A detailed summary of the topic or the main response message")
+    key_metrics: list[str] = Field(description="A list of key metrics, facts, or bullet points related to the query")
