+++ pyproject.toml
@@ -20,6 +20,8 @@ dependencies = [
     "loguru (>=0.7.3,<0.8.0)",
     "fastapi[standard] (>=0.123.0,<0.124.0)",
     "slowapi (>=0.1.9,<0.2.0)",
+    "langchain-community (>=0.4.1,<0.5.0)",
+    "langchain-openai (>=1.2.1,<2.0.0)",
 ]
 
 [project.scripts]
